Langchain/FilmAgent/requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pytest>=7.4.0
ollama>=0.1.0
rich>=13.0.0
